+++ src/templates/blog-post.jsx
@@ -1,20 +1,24 @@
 import React from 'react'
 import Helmet from 'react-helmet'
 import get from 'lodash/get'
+import { Share } from 'react-twitter-widgets'
 
 import Bio from '../components/bio'
 import { rhythm, scale } from '../utils/typography'
+
+import './blog-post.scss'
 
 class BlogPostTemplate extends React.Component {
   render() {
     const post = this.props.data.markdownRemark
     const siteTitle = get(this.props, 'data.site.siteMetadata.title')
+    const postTitle = post.frontmatter.title
 
     return (
       <div>
-        <Helmet title={`${post.frontmatter.title} | ${siteTitle}`} />
-        <h1>{post.frontmatter.title}</h1>
-        <p
+        <Helmet title={`${postTitle} | ${siteTitle}`} />
+        <h1>{postTitle}</h1>
+        <section
           style={{
             ...scale(-1 / 5),
             display: 'block',
@@ -23,8 +27,14 @@ class BlogPostTemplate extends React.Component {
           }}
         >
           {post.frontmatter.date}
-        </p>
+        </section>
         <div dangerouslySetInnerHTML={{ __html: post.html }} />
+        <section className='share-section'>
+          <p className='prompt'>Let's keep the conversation going!</p>
+          <div className='social'>
+            <Share options={{ size: 'small', text: `Check out '${postTitle}'`, via: 'cwpittman' }} />
+          </div>
+        </section>
         <hr
           style={{
             marginBottom: rhythm(1),
